+++ src/components/tab-header-item/__tests__/TabHeaderItemWrapper.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import { ThemeProvider } from 'styled-components';
-import defaultTheme from 'sinoui-components/styles/defaultTheme';
+import { defaultTheme } from '@sinoui/theme';
 import 'jest-styled-components';
 import TabHeaderItemWrapper from '../TabHeaderItemWrapper';