+++ src/dev.d.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 // import original module declarations
 import 'styled-components';
-import { Theme } from 'sinoui-components/styles';
+import { Theme } from '@sinoui/theme';
 
 // and extend them!
 declare module 'styled-components' {